Make bugreport parser compatible with updated metadata tags
With some metadata tags expanded, i.e their full name written beside the id in brackets, make the parser compatible with it to ignore the value in brackets for those expanded tags - so it generates the exact same json it did before.

Additionally update the script to python 3.

Bug: 231590858
Test: Verified that the parser generates the exact same json in a clean
directory as it does with the changes in this branch.

Change-Id: I9f09c07d49527828567b8280f71273022b9587a6